Fix submit button in sample app
The `useIsSubmitting` hook needs to be called from inside a validated form
